fix skill tree hook deps
